upgrade_data/requirements.txt
aiohttp>=3.9.0
asyncio
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
sqlite3; python_version < '3'
